\name{DMQ-package}
\alias{DMQ-package}
\alias{DMQ}
\docType{package}
\title{
Dynamic Multiple Quantiles model in R
}
\description{
The DMQ package allows us to simulate, estimate and forecast using the Dynamic Multiple Quantile (DMQ) model of Catania, L., & Luati, A. (2022). Semiparametric modeling of multiple quantiles. Journal of Econometrics, see \doi{10.1016/j.jeconom.2022.11.002}.
}
\note{
Please cite Catania and Luati (2022) in working papers and published papers that use \code{DMQ}. Use \code{citation("DMQ")}.
}
\author{
Leopoldo Catania [aut,cre], Alessandra Luati [aut]

Maintainer: Leopoldo Catania <leopoldo.catania@econ.au.dk>
}
\references{
Catania, L, and Luati, A. (2023). 
"Semiparametric modeling of multiple quantiles."
Journal of Econometrics
\doi{10.1016/j.jeconom.2022.11.002}.\cr
}
\keyword{ package }
